Refactor ProjectsSection and AboutSection components for improved functionality and organization; implement form handling utility in ContactSection
