+++ src/sections/AboutSection/AboutSoftSkills.jsx
@@ -6,32 +6,14 @@ import { MdOutlineHandshake as CollabIcon } from "react-icons/md";
 import { LuTimer as TimeIcon } from "react-icons/lu";
 import { LuPocketKnife as AdaptIcon } from "react-icons/lu";
 
-const softSkills = [
-  {
-    title: "Communication",
-    description:
-      "I've learned to communicate clearly and effectively with people from different backgrounds and nationalities.",
-    icon: <CommsIcon />,
-  },
-  {
-    title: "Collaboration",
-    description:
-      "I've worked with developers from all over the world, and I've learned to work in a team, manage projects, and meet deadlines.",
-    icon: <CollabIcon />,
-  },
-  {
-    title: "Time Management",
-    description:
-      "I've learned to manage my time effectively and efficiently, especially when working on multiple projects.",
-    icon: <TimeIcon />,
-  },
-  {
-    title: "Adaptability",
-    description:
-      "I've learned to adapt to new technologies, tools, and environments.",
-    icon: <AdaptIcon />,
-  },
-];
+import { softSkills } from "./aboutUtils";
+
+const icons = {
+  'communication': <CommsIcon />,
+  'collaboration': <CollabIcon />,
+  'time management': <TimeIcon />,
+  'adaptability': <AdaptIcon />,
+};
 
 const AboutSoftSkills = () => {
   return (
@@ -45,7 +27,7 @@ const AboutSoftSkills = () => {
       <ul className="about__soft-skills--list grid grid-col grid-gap-2">
         {softSkills.map((skill, index) => (
           <Transition
-            key={index}
+            key={skill.id}
             className="about__soft-skills--list__item p-2"
             elementTag="li"
             trigger="whileInView"
@@ -53,8 +35,9 @@ const AboutSoftSkills = () => {
             duration={1}
             delay={index * 0.2}
           >
-            <h4 className="flex flex-ai-c gap-1">
-              {skill.icon} {skill.title}:
+            <h4 className="flex flex-ai-c gap-1">          
+              {icons[skill.title.toLowerCase()]}
+              {skill.title}
             </h4>
 
             <p>{skill.description}</p>
