+++ src/sections/ProjectsSection/ProjectList.jsx
@@ -10,17 +10,13 @@ const ProjectList = () => {
   const [activeCard, setActiveCard] = useState(isTablet ? 2 : 0);
   const data = ProjectData.slice(0, 5);
 
-  const handleHoverOrFocus = (index) => {
-    setActiveCard(index);
-  };
-
   return (
     <ul className="project__list flex flex-col gap-1 flex_md-row">
       {data.map((item, index) => (
         <ProjectCard
           key={index}
           data={item}
-          onHoverOrFocus={() => handleHoverOrFocus(index)}
+          onHoverOrFocus={() => setActiveCard(index)}
           isActive={activeCard === index}
         />
       ))}
